test(mcp): extract wait_for_files helper
Move duplicated JSONL trace-file polling into WaitHelpers and import it from the MCP tracing, agentic, and HTTP router tests.\n\nVerified with: cd mcp_server && mix test test/ptc_runner_mcp/tracing_test.exs test/ptc_runner_mcp/agentic_test.exs test/ptc_runner_mcp/http_router_test.exs
