+++ mcp_server/test/ptc_runner_mcp/tracing_test.exs
@@ -15,6 +15,8 @@ defmodule PtcRunnerMcp.TracingTest do
   # async: false because TraceConfig + TraceHandler attach process-wide.
   use ExUnit.Case, async: false
 
+  import PtcRunnerMcp.TestSupport.WaitHelpers, only: [wait_for_files: 2]
+
   alias PtcRunnerMcp.{JsonRpc, TraceConfig, TraceHandler}
 
   @program_ok "(+ 1 2)"
@@ -78,29 +80,6 @@ defmodule PtcRunnerMcp.TracingTest do
     |> File.read!()
     |> String.split("\n", trim: true)
     |> Enum.map(&Jason.decode!/1)
-  end
-
-  defp wait_for_files(dir, expected, timeout_ms \\ 1_000) do
-    deadline = System.monotonic_time(:millisecond) + timeout_ms
-    do_wait_for_files(dir, expected, deadline)
-  end
-
-  defp do_wait_for_files(dir, expected, deadline) do
-    files = File.ls!(dir) |> Enum.filter(&String.ends_with?(&1, ".jsonl"))
-
-    cond do
-      length(files) >= expected ->
-        files
-
-      System.monotonic_time(:millisecond) > deadline ->
-        flunk(
-          "expected at least #{expected} jsonl files; got #{length(files)}: #{inspect(files)}"
-        )
-
-      true ->
-        Process.sleep(20)
-        do_wait_for_files(dir, expected, deadline)
-    end
   end
 
   # ----------------------------------------------------------------
